Fold layout conversion for TMEM Store to fix perf drop for flex attn (#8353)
This forward propagation part of remove layout conversion pass makes
this

```
%acc_316 = arith.truncf %p_308 : tensor<128x64xf32, #blocked> to tensor<128x64xbf16, #blocked> loc(#loc247)
      ttng.tmem_store %acc_316, %acc_35, %true : tensor<128x64xbf16, #blocked> -> !ttg.memdesc<128x64xbf16, #tmem2, #ttng.tensor_memory, mutable> loc(#loc228)
```

become this

```
%acc_315 = arith.truncf %p_307 : tensor<128x64xf32, #linear> to tensor<128x64xbf16, #linear> loc(#loc247)
      %acc_316 = ttg.convert_layout %acc_315 : tensor<128x64xbf16, #linear> -> tensor<128x64xbf16, #blocked3> loc(#loc247)
      ttng.tmem_store %acc_316, %acc_35, %true : tensor<128x64xbf16, #blocked3> -> !ttg.memdesc<128x64xbf16, #tmem2, #ttng.tensor_memory, mutable> loc(#loc228)
```

The `linear` layout comes from `TMemLoadReducePattern` for `tmem_load`,
so we'd like to respect that if possible.

This commit folds the convert op into TMEMStoreOp if it's compatible,
getting rid of an unnecessary convert_layout op in the main loop for the
flex attn kernel in #8328.

This gets flex attention kernel in #8328 to achieve 499 tflops on 750W
B200. (The regression was from 435 to 359 tflops)

TTGIR files:
- before:
https://gist.github.com/pchen7e2/f47639a166fbf4ea8b01a723ec3899dd
- after:
https://gist.github.com/pchen7e2/fd4eb279d8927efb48711511582443ca