[Gluon] Require warp_specialize default_args and worker_args be tuples (#8368)
I'm not a fan of using `isinstance` here as it leads to an
inconsistency. Say we want to pass a single argument `x` that happens to
be a tuple of tensors, then that has to be passed as `(x,)` or otherwise
the tuple will be interpreted as a list of arguments to pass to the
worker function.

The original PR (#8269) justified this as improving the error message,
so I also add a validation check with a better error message if you
forget to pass a tuple.