[BACKEND] Fix swizzling=0 support for matrix descriptors and generalize wgmma lowering (#8378)
We fix a bug we had in the matrix descriptor lowering. We also add
comprehensive gluon tests for the `swizzling=0` case. To do so, we allow
`wgmma` to accept `SharedLinearLayout` when it takes a shared memory
argument.

While testing all the cases for `swizzling=0` we found a `ptxas` bug in
some of them.
For now, we disable `ptxas` optimisations in some of the test cases to
work around this.