[NFC] Python3.13 fix (#8403)
Python 3.13.8 was released yesterday but introduced a
`inspect.getsourceline` regression which causes `test_scan2d` to fail
when using the `get_first_element` op. We can work around it by shifting
the comment out from under the decorator call.

There might be other tests that hit this issue but this is the only one
flagged by our internal CI.

https://github.com/python/cpython/issues/139783